Replace 'short' with 'int16_t'
More clearly expresses the intend (and might also make the code more portable).
Suggested by clang-tidy [google-runtime-int].

In a few places where the type really didn't matter I replaced 'short' with
'int'.

+++ src/events/InputEventGenerator.hh
@@ -88,7 +88,7 @@ private:
 	void setNewOsdControlButtonState(
 		unsigned newState, const EventPtr& origEvent);
 	void triggerOsdControlEventsFromJoystickAxisMotion(
-		unsigned axis, short value, const EventPtr& origEvent);
+		unsigned axis, int value, const EventPtr& origEvent);
 	void triggerOsdControlEventsFromJoystickHat(
 		int value, const EventPtr& origEvent);
 	void osdControlChangeButton(
